Reimplement Map component and put utils functions back into component
